// web/src/api/generated/model/deskCopy.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Campus Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface DeskCopy {
  copy_id: string;
  barcode: string;
  title: string;
  authors: string[];
  edition: string | null;
  call_number: string | null;
  status: string;
  on_loan_to: string | null;
  held_for_person_id: string | null;
  held_for: string | null;
}